requirements.txt
# Calm Turtle — runtime dependencies
alpaca-py>=0.30
pyyaml>=6.0
jsonschema>=4.21
requests>=2.31
python-dateutil>=2.9
pytz>=2024.1
pandas>=2.2
numpy>=1.26

# Dev / tests
pytest>=8.0
ruff>=0.5
